requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
djangorestframework==3.14.0
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.1.0
pyfiglet==0.8.post1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
